Create Blockchain Technology
Within GEDE we opened a discussion about the relevance of BCT in science. The discussions are controversial. We will draft and upload a first document soon that mirrors some of the discussions that happened during the last months. We also upload a few basic papers and welcome a critical debate.